wf tabs now highlight correctly, moved some controllers to sidetabs controllers
