+++ app/controllers/dashboard_controller.rb
@@ -1,8 +1,14 @@
 class DashboardController < ApplicationController
   def index
-
     @workflows = Workflow.limit(20).reverse_order
-    @workflow = false
+    @workflow = false # For workflow-content to display empty
 
+    render 'dashboard/index'
+  end
+
+  def my_tasks
+    @my_tasks = current_user.tasks
+
+    render 'my_tasks/index'
   end
 end
