+++ app/javascript/controllers/world_controller.js
@@ -2,7 +2,7 @@ import { Controller } from "stimulus"
 import { csrfToken } from "@rails/ujs"
 
 export default class extends Controller {
-  static targets = ['contentmain', 'workflowslist']
+  static targets = ['contentmain']
 
   connect() {
     console.log('world controller connected')
