+++ app/javascript/stylesheets/application.scss
@@ -9,16 +9,16 @@
 
   // Buttons
   .t-btn-outline {
-    @apply w-full border border-purple-400 rounded-lg p-2 flex items-center hover:bg-purple-400 hover:text-white duration-100 ;
+    @apply w-full border border-indigo-400 rounded-lg p-2 flex items-center hover:bg-indigo-400 hover:text-white duration-100 ;
   }
   .t-workflow-btn-active {
-    @apply px-4 py-2 my-1 text-sm rounded-lg bg-purple-100 duration-100 ;
+    @apply px-4 py-2 my-1 text-sm rounded-lg bg-indigo-100 duration-100 ;
   }
   .t-workflow-btn {
     @apply px-4 py-2 my-1 text-sm rounded-lg hover:bg-gray-100 duration-100 ;
   }
   .t-tab-active {
-    @apply text-sm font-medium text-purple-600 pb-1 border-b-2 border-purple-600 ;
+    @apply text-sm font-medium text-indigo-600 pb-1 border-b-2 border-indigo-600 ;
   }
   .t-tab {
     @apply text-sm font-medium text-gray-400 pb-1 border-b-2 border-white ;
