+++ app/models/user.rb
@@ -16,6 +16,7 @@ class User < ApplicationRecord
 
   has_many :task_members, dependent: :destroy
   has_many :tasks, through: :task_members
+
   # has_many :item_members, dependent: :destroy
   # has_many :items, through: :item_members
