--- app/views/items/index.html.erb
@@ -1,9 +1,0 @@
-<%# <div data-controller="item">
-  <ol class='list-none'>
-    <div data-item-target="list">
-    </div>
-  </ol>
-  <div data-item-target="newform">
-  </div>
-  <%= render "items/itemadd" %>
-</div> %>
