+++ app/views/my_tasks/_task_tab.html.erb
@@ -1,4 +1,4 @@
 <%# FOR AJAX ONLY %>
-<div class="t-workflow-btn-active">
+<div data-action="click->world#selectWorkflow" class="t-workflow-btn-active">
   <%= link_to workflow.title, workflow_path(workflow), id: 'workflow', remote: true %>
 </div>
