+++ app/views/my_tasks/_tasks_tabs.html.erb
@@ -1,0 +1,14 @@
+<% my_tasks.each do |task| %>
+
+  <% if workflow == @workflow %>
+    <div data-workflow-id=<%= workflow.id %> data-controller="tabs" class="t-workflow-btn-active flex items-center justify-between">
+  <% else %>
+    <div data-workflow-id=<%= workflow.id %> data-controller="tabs" class="t-workflow-btn flex items-center justify-between">
+  <% end %>
+
+    <%= link_to workflow_path(workflow), class: "flex-grow" do %>
+        <%= workflow.title %>
+    <% end %>
+  </div>
+
+<% end %>
