+++ app/views/my_tasks/index.html.erb
@@ -1,0 +1,20 @@
+<div class="flex">
+
+  <div class="w-72 p-5 shadow-lg fixed h-screen">
+    <%= link_to '🅗Taskete', dashboard_path, class: 'font-medium pb-3' %>
+    <%= render "my_tasks/my_tasks_l" %>
+  </div>
+  <div class="w-72"></div> <%# ghost sidebar to push content %>
+
+
+  <div class="flex-grow pt-4 pl-10">
+    <div id="my_tasks-content">
+      <%# <%= render "my_tasks/my_tasks_content", locals: {} %>
+    </div>
+  </div>
+
+  <div class="w-60 pt-2 px-5">
+    <%= render 'shared/sidebar_r' %>
+  </div>
+
+</div>
