+++ app/views/tasks/_task.html.erb
@@ -17,6 +17,7 @@
 
     </div>
 
+    <%# this holds/inserts task_content %>
     <div class="flex hidden" data-item-target="taskcontent" data-toggle-target="hide" >
     </div>
